refactor component to shared directory
